+++ pyNastran/bdf/cards/elements/solid.py
@@ -2447,7 +2447,7 @@ class CPYRAM13(SolidElement):
             the BDF object
         """
         msg = ', which is required by CPYRAM eid=%s' % self.eid
-        self.nodes_ref = model.Nodes(self.nodes, msg=msg)
+        self.nodes_ref = model.Nodes(self.nodes[:5], msg=msg) + model.EmptyNodes(self.nodes[5:], msg=msg)
         self.pid_ref = model.safe_property(self.pid, self.eid, xref_errors, msg=msg)
 
     @property